build(macos): Don't bake in libobs rpath
CMake will generate -Wl,-rpath instructions otherwise.
These in turn make dyld generate LC_RPATH entries with the (temporary)
libobs path in the binary. dyld will therefore try and load dependent
libraries from there.
This is not usually a problem, as a given user will most likely not have
the path in the LC_RPATH instruction exist. In cases where it does
(e.g. for people that build obs-websocket themselves) it may lead to
library loading incosnistencies.
At any rate it leads to artifacts of the build environment retained in the
binary, which should be avoided.