CI(macos): remove absolute path from Qt lib ids
The packaging script left the library ID of the Qt libraries that
are shipped and integrated into OBS.app as is (i.e. /usr/local/Cellar...).
This may lead to library loading inconsitencies.
Furthermore, all the Qt libraries already present in OBS.app have their
ID starting with @rpath as well. This commit aligns the ones here with
what is observed in OBS.