CI(macos): only copy Qt libs if they don't exist
This is mainly a fix for users that try to build obs-websocket themselves.
From a previous (failed) attempt, the Qt libraries may already exist in
the build directory, however marked as read-only. Therefore, cp cannot
overwrite them, even though it would be perfectly fine to do so.
This commit makes it so the copy is not attempted if the libraries exist.