CI(macos): When finding Qt version, sort by version
This fix is intended mainly for self-builders. If there are old versions
of Qt around in /usr/local/Cellar/qt, probably because the user has not
run "brew cleanup" in a while, the package script would erroneously try
and replace the wrong path in the Qt binaries.
This may lead to library loading problems, as some paths would not be
replaced with their @rpath versions.