utils: Add `inputKindCaps` field to input list array
I debated adding this to the input kinds array, where this field's
value truly comes from, but the value of that return is
`Array<String>`, so we don't really have any options without bumping
the current RPC version.

I am thinking of revisiting some aspects of the obs-websocket request
objectives, and bumping the RPC version. I would like to take a more
REST-style approach to reading/modification of OBS state.