eventhandler: Fix possible crash on OBS start in Studio Mode
When OBS Studio is started in Studio Mode, if a scene collection was not
successfully found and loaded, the OBS Frontend API would, prior to
https://github.com/obsproject/obs-studio/commit/577e3504d59f0d8e488ada7a05295d1d8267f095,
emit OBS_FRONTEND_EVENT_SCENE_CHANGED and
OBS_FRONTEND_EVENT_PREVIEW_SCENE_CHANGED anyway. This would result in
currentScene here being a nullptr, causing a crash.

For completeness, check for nullptr here so that we cannot crash if
obs_frontend_get_current_scene() returns a nullptr.